requirements.txt
pytube==12.1.0
librosa==0.8.1
numpy==1.22.4
scipy==1.10.1
ffmpeg-python==0.2.0
spleeter==2.3.2
pydub==0.25.1
wave==0.0.2
pyaudio==0.2.14